refactor aise directory structure
